implement move style file
