+++ src/styles/NoteInput.css
