+++ src/styles/NoteItem.css
