+++ src/styles/NotesListArea.css
